fix latitude and longitude magnitude

+++ GroundSegment/src/radio_link/simulator/handlers/agu_send_telemetry_handler.h
@@ -22,8 +22,8 @@ namespace radiolink
     private:
         UavModel* m_model;
         uint64_t generateMask();
-        float m_lat;
-        float m_lon;
+        uint32_t m_lat;
+        uint32_t m_lon;
     };
 }
